+++ brew_shareapi/serializers/__init__.py
@@ -1,0 +1,3 @@
+from .entry import EntrySerializer
+from .coffee import CoffeeSerializer
+from .brew_method import MethodSerializer
