+++ brew_shareapi/serializers/brew_method.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from brew_shareapi.models import BrewMethod
+
+class MethodSerializer(serializers.ModelSerializer):
+    """JSON serializer for brew methods"""
+    class Meta:
+        model = BrewMethod
+        fields = ('id', 'name','method_image', 'website', )
