+++ brew_shareapi/serializers/brewer.py
@@ -1,0 +1,12 @@
+from rest_framework import serializers
+from django.contrib.auth.models import User
+from brew_shareapi.models import Brewer
+
+class UserSerializer(serializers.ModelSerializer):
+    """JSON serializer for brewer's related Django user"""
+    class Meta:
+        model = User
+        fields = ('first_name', 'last_name', 'username', )
+
+class BrewerSerializer(serializers.ModelSerializer):
+    """JSON serializer for """
